删除服务端模式的TLS支持（后续会换成TLS.13）
检测字符串形式的数字

重构EnumHelper

utf8mb4 again

arraycache bugfix

asm性能优化
按需解析UTF-8字符串
按需生成常量set
...

lazy utf creation

forEachByte(IntUnaryOperator)

progressbar speed fix

Cipher malfunction on AEAD

DynByteBuf expand before

MyPassIs v2

misc bug fix

+++ roj/asm/visitor/SwitchSegment.java
@@ -63,7 +63,7 @@ public final class SwitchSegment extends Segment {
 		o.put(code);
 
 		int pad = (4 - ((self+1) & 3)) & 3;
-		while (pad-- > 0) o.put((byte) 0);
+		while (pad-- > 0) o.put(0);
 
 		List<SwitchEntry> m = targets;
 		if (code == TABLESWITCH) {
